+++ src/OpenSimCreator/UI/MeshWarper/MeshWarpingTabInputMeshPanel.hpp
@@ -515,6 +515,10 @@ namespace osc
                 {
                     ActionSaveLandmarksToCSV(m_State->getScratch(), m_DocumentIdentifier);
                 }
+                if (ImGui::MenuItem("Landmarks to CSV (only positions, no header or names)"))
+                {
+                    ActionSaveLandmarksToCSV(m_State->getScratch(), m_DocumentIdentifier, TPSDocumentCSVFlags::NoHeader | TPSDocumentCSVFlags::NoNames);
+                }
                 ImGui::EndPopup();
             }
         }