+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented here. The format is based
 - Fixed a bug where user-defined muscle points could not be selected if the user was rendering muscles
   in tendons+fibers mode (#706)
 - Landmarks in the mesh warper UI can now be renamed and repositioned via a right-click context menu (#807)
+- The mesh warper's CSV exporter for landmarks now includes a header row and name column (#807)
+  - The previous functionality (just numbers) is exposed as "(only positions, no header or names)"
 
 
 ## [0.5.5] - 2023/11/21